splitting program in separate assignments

+++ Helper/NearestNeighbours.cs
@@ -6,7 +6,7 @@ namespace INFDTA01_1.Helper
 {
     public static class NearestNeighbours
     {
-        public static Dictionary<int, double> Compute(SortedDictionary<int, double> similarities, float threshold, int limit = 3)
+        public static Dictionary<int, double> Compute(SortedDictionary<int, double> similarities, double threshold, int limit = 3)
         {
             // remove threshold exceedings
             var toRemove = similarities.Where(similarity => similarity.Value < threshold)
